Fixed many bugs

+++ src/UC_Quantum_Lab/layout.py
@@ -1,6 +1,5 @@
 from . import _states, _circs, _hists, _layout_file, _master_show, _config_dir, _trigger_file
 from ._src import _trigger
-from atexit import register
 import os, json, platform
 
 _layout = {}
@@ -139,12 +138,9 @@ def _layout_at_exit():
         _run()
         _trigger()
 
-register(_layout_at_exit)
-
 def _exit():
-    for item in os.listdir(_config_dir):
-        # deletes png html or the trigger file from the config dir
-        if item == _trigger_file:
-            os.remove(os.path.join(_config_dir, item))
-
-register(_exit)
+    if _config_dir in os.listdir():
+        for item in os.listdir(_config_dir):
+            # deletes png html or the trigger file from the config dir
+            if item == _trigger_file:
+                os.remove(os.path.join(_config_dir, item))
